MAINT: Move `empty_like` to `np.core.multiarray`
